final code re-structuring
