+++ requirements.txt
@@ -6,5 +6,4 @@ numpy==1.26.4
 pandas
 scikit_learn
 spacy
-textblob
 wordcloud
